+++ custom_components/whatsapp/const.py
@@ -21,6 +21,7 @@ ADDON_FALLBACK_HOSTS: Final = (
 
 SERVICE_SEND_MESSAGE: Final = "send_message"
 SERVICE_CHECK_NUMBER: Final = "check_number"
+SERVICE_GET_GROUP_INFO: Final = "get_group_info"
 SERVICE_SET_STATUS: Final = "set_status"
 SERVICE_PRESENCE_SUBSCRIBE: Final = "presence_subscribe"
 SERVICE_SEND_PRESENCE_UPDATE: Final = "send_presence_update"
@@ -28,6 +29,7 @@ SERVICE_SEND_INFINITY_PRESENCE_UPDATE: Final = "send_infinity_presence_update"
 SERVICE_READ_MESSAGES: Final = "read_messages"
 
 CAPABILITY_CHECK_NUMBER: Final = SERVICE_CHECK_NUMBER
+CAPABILITY_GET_GROUP_INFO: Final = SERVICE_GET_GROUP_INFO
 
 ATTR_BODY: Final = "body"
 ATTR_CLIENT_ID: Final = "clientId"